requirements.txt
Flask
groq
google-generativeai